Resumen reportes

+++ main.js
@@ -176,9 +176,7 @@ function mostrarCategoria() {
     editButton.addEventListener("click", function () {
       mostrarSeccion(editCategoria);
       const nombreCategoriaExistente = categorias[i];
-      const editarCategoriaInput = document.getElementById(
-        "editar-categoria-nombre"
-      );
+      const editarCategoriaInput = document.getElementById("editar-categoria-nombre");
       editarCategoriaInput.value = nombreCategoriaExistente;
 
       const confirmarEditar = document.getElementById("confirm-edit");
@@ -188,8 +186,21 @@ function mostrarCategoria() {
         if (nuevoNombre !== "") {
           categorias[i] = nuevoNombre;
           localStorage.setItem("categorias", JSON.stringify(categorias));
-          mostrarCategoria();
-          mostrarSeccion(categoriasSec);
+
+           // Actualizar las categorías en las operaciones
+           let operaciones = JSON.parse(localStorage.getItem("operaciones")) || [];
+           operaciones = operaciones.map(op => {
+             if (op.categoriaOperacion === nombreCategoriaExistente) {
+               op.categoriaOperacion = nuevoNombre;
+             }
+             return op;
+           });
+           localStorage.setItem("operaciones", JSON.stringify(operaciones));
+ 
+           mostrarCategoria();
+           mostrarSeccion(categoriasSec);
+           generarReporte();
+           mostrarOperaciones();
         }
       });
     });
@@ -205,22 +216,27 @@ function mostrarCategoria() {
 
     deleteButton.addEventListener("click", function () {
       const categoriaAEliminar = categorias[i];
-
-      // Eliminar la categoría
-      categorias = categorias.filter(
-        (categoria) => categoria !== categoriaAEliminar
-      );
-      localStorage.setItem("categorias", JSON.stringify(categorias));
-
-      // Eliminar las operaciones relacionadas
-      let operaciones = JSON.parse(localStorage.getItem("operaciones")) || [];
-      operaciones = operaciones.filter(
-        (operacion) => operacion.categoriaOperacion !== categoriaAEliminar
-      );
-      localStorage.setItem("operaciones", JSON.stringify(operaciones));
-
-      mostrarOperaciones();
-      mostrarCategoria();
+    
+      // Mostrar ventana de confirmación
+      const confirmacion = confirm(`¿Estás seguro de que quieres eliminar la categoría "${categoriaAEliminar}"?`);
+    
+      if (confirmacion) {
+        // Eliminar la categoría
+        categorias = categorias.filter(
+          (categoria) => categoria !== categoriaAEliminar
+        );
+        localStorage.setItem("categorias", JSON.stringify(categorias));
+    
+        // Eliminar las operaciones relacionadas
+        let operaciones = JSON.parse(localStorage.getItem("operaciones")) || [];
+        operaciones = operaciones.filter(
+          (operacion) => operacion.categoriaOperacion !== categoriaAEliminar
+        );
+        localStorage.setItem("operaciones", JSON.stringify(operaciones));
+    
+        mostrarOperaciones();
+        mostrarCategoria();
+      }
     });
 
     // Añadir contenedor de botones al contenedor principal
@@ -419,12 +435,8 @@ function mostrarOperaciones(operaciones = null) {
           document.getElementById("editar-op-btn");
 
         confirmarEditarOperacion.addEventListener("click", function () {
-          const nuevaDescripcion = document.getElementById(
-            "descripcion-edit-op"
-          ).value;
-          const nuevoMonto = parseFloat(
-            document.getElementById("monto-edit-op").value
-          );
+          const nuevaDescripcion = document.getElementById("descripcion-edit-op").value;
+          const nuevoMonto = parseFloat(document.getElementById("monto-edit-op").value);
           const nuevoTipo = document.getElementById("tipo-edit-op").value;
           const nuevaCategoria = document.getElementById(
             "select-categoria-edit"
@@ -442,6 +454,7 @@ function mostrarOperaciones(operaciones = null) {
           localStorage.setItem("operaciones", JSON.stringify(operaciones));
           mostrarOperaciones();
           mostrarSeccion(balance);
+          generarReporte();
         });
       });
 
@@ -450,10 +463,23 @@ function mostrarOperaciones(operaciones = null) {
       botonEliminar.textContent = "Eliminar";
       botonEliminar.className = "text-red-500 hover:underline text-xs";
       botonEliminar.addEventListener("click", function () {
+
+        // Obtener la descripción y la categoría de la operación
+      const descripcionOperacion = operaciones[index].descripcionOperacion;
+      const categoriaOperacion = operaciones[index].categoriaOperacion;
+
+      // Mostrar mensaje de confirmación con la descripción y la categoría
+      const confirmar = confirm(`¿Estás seguro de que quieres eliminar la operación?\n\nDescripción: ${descripcionOperacion}\nCategoría: ${categoriaOperacion}`);
+
+      if (confirmar) {
+        // Si el usuario confirma, eliminar la operación
         operaciones.splice(index, 1);
         localStorage.setItem("operaciones", JSON.stringify(operaciones));
         mostrarOperaciones();
-      });
+        generarReporte();
+        actualizarResumen(operaciones); 
+      }
+      }); 
 
       acciones.appendChild(botonEditar);
       acciones.appendChild(botonEliminar);
@@ -619,3 +645,211 @@ function actualizarResumen(operaciones) {
   totalGeneral.textContent = `$${totalSigno}`;
   totalGeneral.className = `text-xl font-bold ${totalColor}`;
 }
+
+//----------------------------------  REPORTES --------------------------------------------------
+
+function generarReporte() {
+  const operaciones = JSON.parse(localStorage.getItem("operaciones")) || [];
+  const categorias = JSON.parse(localStorage.getItem("categorias")) || [];
+
+  const sinReportes = document.getElementById("sin-reportes");
+  const contenedorReporte = document.getElementById("contenedor-reporte");
+
+  // Si no hay operaciones, muestra el mensaje de "Operaciones insuficientes"
+  if (operaciones.length === 0) {
+    sinReportes.classList.remove("hidden");
+    contenedorReporte.innerHTML = "";
+    return; 
+  } else {
+    sinReportes.classList.add("hidden"); 
+  }
+
+
+  let resumen = {
+    categoriaMayorGanancia: { nombre: "", ganancia: 0 },
+    categoriaMayorGasto: { nombre: "", gasto: 0 },
+    categoriaMayorBalance: { nombre: "", balance: 0 },
+    mesMayorGanancia: { nombre: "", ganancia: 0 },
+    mesMayorGasto: { nombre: "", gasto: 0 },
+  };
+
+  const totalesPorCategoria = {};
+  const totalesPorMes = {};
+
+  // Inicializar los objetos para los totales por categoría y mes
+  categorias.forEach((categoria) => {
+    totalesPorCategoria[categoria] = { ganancia: 0, gasto: 0, balance: 0 };
+  });
+
+  operaciones.forEach((operacion) => {
+    const { categoriaOperacion, fechaOperacion, montoOperacion, tipoOperacion } = operacion;
+    const [anio, mes] = fechaOperacion.split("-");
+
+    if (!totalesPorMes[`${anio}-${mes}`]) {
+      totalesPorMes[`${anio}-${mes}`] = { ganancia: 0, gasto: 0, balance: 0 };
+    }
+
+    if (tipoOperacion === "GANANCIA") {
+      totalesPorCategoria[categoriaOperacion].ganancia += montoOperacion;
+      totalesPorMes[`${anio}-${mes}`].ganancia += montoOperacion;
+    } else if (tipoOperacion === "GASTO") {
+      totalesPorCategoria[categoriaOperacion].gasto += montoOperacion;
+      totalesPorMes[`${anio}-${mes}`].gasto += montoOperacion;
+    }
+
+    totalesPorCategoria[categoriaOperacion].balance =
+      totalesPorCategoria[categoriaOperacion].ganancia -
+      totalesPorCategoria[categoriaOperacion].gasto;
+
+    totalesPorMes[`${anio}-${mes}`].balance =
+      totalesPorMes[`${anio}-${mes}`].ganancia - totalesPorMes[`${anio}-${mes}`].gasto;
+  });
+
+  // Calcular los máximos para el resumen
+  Object.entries(totalesPorCategoria).forEach(([categoria, totales]) => {
+    if (totales.ganancia > resumen.categoriaMayorGanancia.ganancia) {
+      resumen.categoriaMayorGanancia = { nombre: categoria, ganancia: totales.ganancia };
+    }
+    if (totales.gasto > resumen.categoriaMayorGasto.gasto) {
+      resumen.categoriaMayorGasto = { nombre: categoria, gasto: totales.gasto };
+    }
+    if (totales.balance > resumen.categoriaMayorBalance.balance) {
+      resumen.categoriaMayorBalance = { nombre: categoria, balance: totales.balance };
+    }
+  });
+
+  Object.entries(totalesPorMes).forEach(([mes, totales]) => {
+    if (totales.ganancia > resumen.mesMayorGanancia.ganancia) {
+      resumen.mesMayorGanancia = { nombre: mes, ganancia: totales.ganancia };
+    }
+    if (totales.gasto > resumen.mesMayorGasto.gasto) {
+      resumen.mesMayorGasto = { nombre: mes, gasto: totales.gasto };
+    }
+  });
+
+  mostrarReporte(resumen, totalesPorCategoria, totalesPorMes);
+}
+
+function mostrarReporte(resumen, totalesPorCategoria, totalesPorMes) {
+  const contenedorReporte = document.getElementById("contenedor-reporte");
+  contenedorReporte.innerHTML = "";
+
+  const resumenHtml = `
+  <div class="p-6 bg-white border border-gray-200 rounded-lg shadow-md mb-8 mt-6">
+      <h3 class="text-xl font-semibold mb-4">Resumen</h3>
+      <div class="grid grid-cols-1 md:grid-cols-2 gap-y-2 gap-x-4 text-sm">
+          <p class="font-semibold">Categoría con mayor ganancia</p>
+          <div class="border-b border-gray-300 md:border-b-0 flex items-center justify-between space-x-2">
+              <span class="italic w-auto text-center font-medium text-purple-600">${resumen.categoriaMayorGanancia.nombre}</span>
+              <span class="text-green-600 font-bold ml-4">+$${resumen.categoriaMayorGanancia.ganancia}</span>
+          </div>
+          <p class="font-semibold">Categoría con mayor gasto</p>
+          <div class="border-b border-gray-300 md:border-b-0 flex items-center justify-between space-x-2">
+              <span class="italic w-auto text-center font-medium text-purple-600">${resumen.categoriaMayorGasto.nombre}</span>
+              <span class="text-red-600 font-bold ml-4">-$${resumen.categoriaMayorGasto.gasto}</span>
+          </div>
+          <p class="font-semibold">Categoría con mayor balance</p>
+          <div class="border-b border-gray-300 md:border-b-0 flex items-center justify-between space-x-2">
+              <span class="italic w-auto text-center font-medium text-purple-600">${resumen.categoriaMayorBalance.nombre}</span>
+              <span class="font-bold ml-4">$${resumen.categoriaMayorBalance.balance}</span>
+          </div>
+          <p class="font-semibold">Mes con mayor ganancia</p>
+          <div class="border-b border-gray-300 md:border-b-0 flex items-center justify-between space-x-2">
+              <span class="text-gray-500">${getMesAnio(resumen.mesMayorGanancia.nombre)}</span>
+              <span class="text-green-600 font-bold ml-4">+$${resumen.mesMayorGanancia.ganancia}</span>
+          </div>
+          <p class="font-semibold">Mes con mayor gasto</p>
+          <div class="border-b border-gray-300 md:border-b-0 flex items-center justify-between space-x-2">
+              <span class="text-gray-500">${getMesAnio(resumen.mesMayorGasto.nombre)}</span>
+              <span class="text-red-600 font-bold ml-4">-$${resumen.mesMayorGasto.gasto}</span>
+          </div>
+      </div>
+  </div>
+  `;
+  // Filtrar las categorías que tienen operaciones (ganancias o gastos)
+  const categoriasConOperaciones = Object.keys(totalesPorCategoria).filter(
+    (categoria) =>
+      totalesPorCategoria[categoria].ganancia > 0 || totalesPorCategoria[categoria].gasto > 0
+  );
+
+
+  let totalesCategoriaHtml = `
+  <h3 class="text-xl font-semibold mb-4">Totales por categorías</h3>
+  <div class="overflow-x-auto mb-8">
+      <table class="min-w-full border border-gray-200">
+          <thead>
+              <tr class="bg-purple-500 text-white">
+                  <th class="border px-4 py-2 text-left">Categoría</th>
+                  <th class="border px-4 py-2 text-left">Ganancias</th>
+                  <th class="border px-4 py-2 text-left">Gastos</th>
+                  <th class="border px-4 py-2 text-left">Balance</th>
+              </tr>
+          </thead>
+          <tbody class="divide-y divide-gray-100">
+  `;
+
+  // Mostrar solo las categorías filtradas
+  categoriasConOperaciones.forEach((categoria) => {
+    const totales = totalesPorCategoria[categoria];
+    totalesCategoriaHtml += `
+      <tr>
+          <td class="border px-4 py-2">${categoria}</td>
+          <td class="border px-4 py-2 text-green-600">+$${totales.ganancia}</td>
+          <td class="border px-4 py-2 text-red-600">-$${totales.gasto}</td>
+          <td class="border px-4 py-2">${totales.balance >= 0 ? '$' + totales.balance : '-$' + Math.abs(totales.balance)}</td>
+      </tr>
+    `;
+  });
+
+  totalesCategoriaHtml += `
+          </tbody>
+      </table>
+  </div>
+  `;
+
+  let totalesMesHtml = `
+  <h3 class="text-xl font-semibold mb-4">Totales por mes</h3>
+  <div class="overflow-x-auto">
+      <table class="min-w-full border border-gray-200">
+          <thead>
+              <tr class="bg-purple-500 text-white">
+                  <th class="border px-4 py-2 text-left">Mes</th>
+                  <th class="border px-4 py-2 text-left">Ganancias</th>
+                  <th class="border px-4 py-2 text-left">Gastos</th>
+                  <th class="border px-4 py-2 text-left">Balance</th>
+              </tr>
+          </thead>
+          <tbody class="divide-y divide-gray-100">
+  `;
+
+  Object.entries(totalesPorMes).forEach(([mes, totales]) => {
+      const mesAnio = getMesAnio(mes);
+
+      totalesMesHtml += `
+      <tr>
+          <td class="border px-4 py-2">${mesAnio}</td>
+          <td class="border px-4 py-2 text-green-600">+$${totales.ganancia}</td>
+          <td class="border px-4 py-2 text-red-600">-$${totales.gasto}</td>
+          <td class="border px-4 py-2">${totales.balance >= 0 ? '$' + totales.balance : '-$' + Math.abs(totales.balance)}</td>
+      </tr>
+      `;
+  });
+
+  totalesMesHtml += `
+          </tbody>
+      </table>
+  </div>
+  `;
+
+  contenedorReporte.innerHTML = resumenHtml + totalesCategoriaHtml + totalesMesHtml;
+}
+
+function getMesAnio(mes) {
+  const [anio, mesNum] = mes.split("-");
+  const mesNombre = new Date(anio, mesNum - 1).toLocaleString('es-ES', { month: 'long' });
+  return `${mesNombre} ${anio}`;
+}
+
+
+document.addEventListener("DOMContentLoaded", generarReporte);
+
